Updates to structure, added some comments
Mostly to test github
