+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Auto-Any-StraightAhead.java
@@ -4,8 +4,8 @@ import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-@Autonomous(name = "Red Right Autonomous", group = "Apex Robotics 3916")
-public class RedRightAuto extends OpMode {
+@Autonomous(name = "Straight Ahead Autonomous", group = "Apex Robotics 3916")
+public class StraightAheadAuto extends OpMode {
     private Robot bot = new Robot();
     private ElapsedTime runtime = new ElapsedTime();
     private String timeOfCompletion;
@@ -15,7 +15,7 @@ public class RedRightAuto extends OpMode {
 
     @Override
     public void init() {
-        bot.init(hardwareMap, Robot.DriveType.Mechanum);
+        bot.init(hardwareMap, Robot.DriveType.Mecanum);
     }
 
     @Override
@@ -34,31 +34,10 @@ public class RedRightAuto extends OpMode {
             // even numbers for performing an action while odd values for state are used for waiting, seen in default case
             case 0:
                 // drive forward
-                new Thread(new Tasker("drive,0,0.5,1500")).start();
+                new Thread(new Tasker("drive,0,0.5,1800")).start();
                 state++;
-
                 break;
             case 2:
-                // turn right
-                new Thread(new Tasker("turn,1,500")).start();
-                state++;
-                break;
-            case 4:
-                // drive forward
-                new Thread(new Tasker("drive,0,0.5,3000")).start();
-                state++;
-                break;
-            case 6:
-                // turn left
-                new Thread(new Tasker("turn,0,500")).start();
-                state++;
-                break;
-            case 8:
-                // drive forward
-                new Thread(new Tasker("drive,0,0.5,500")).start();
-                state++;
-                break;
-            case 10:
                 // autonomous is done as robot has parked
                 telemetryMsg = "autonomous completed in " + runtime.toString() + " seconds";
                 break;
