+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Auto-Blue-Foundation.java
@@ -4,14 +4,8 @@ import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-import org.firstinspires.ftc.robotcore.external.navigation.DistanceUnit;
-
-import java.util.Arrays;
-import java.util.ArrayList;
-import java.util.List;
-
-@Autonomous(name = "Test Autonomous", group = "Apex Robotics 3916")
-public class AutoCode extends OpMode {
+@Autonomous(name = "Base Plate BLUE Autonomous", group = "Apex Robotics 3916")
+public class TestAutoCode extends OpMode {
     private Robot bot = new Robot();
     private ElapsedTime runtime = new ElapsedTime();
     private String timeOfCompletion;
@@ -21,7 +15,7 @@ public class AutoCode extends OpMode {
 
     @Override
     public void init() {
-        bot.init(hardwareMap, Robot.DriveType.Mechanum);
+        bot.init(hardwareMap, Robot.DriveType.Mecanum);
     }
 
     @Override
@@ -39,32 +33,58 @@ public class AutoCode extends OpMode {
         switch (state) {
             // even numbers for performing an action while odd values for state are used for waiting, seen in default case
             case 0:
-                // drive forward
-                new Thread(new Tasker("drive,0,1,3000")).start();
+                // drive backwards
+                new Thread(new Tasker("drive,0,1,1100")).start();
                 state++;
-
                 break;
             case 2:
-                //turn for a bit
-                new Thread(new Tasker("turn,0,1000"));
+                // bring down base grabby claw
+                new Thread(new Tasker("BP claw down,700")).start(); // command, pause
                 state++;
                 break;
             case 4:
-                //go forward but sideways
-                new Thread(new Tasker("drive,0,1,3000"));
+                // drive forward
+                new Thread(new Tasker("drive,0,-1,1100")).start();
                 state++;
                 break;
             case 6:
-                // autonomous is done as robot has parked
-                telemetryMsg = "autonomous completed in " + runtime.toString() + " seconds";
+                // turn left
+                new Thread(new Tasker("slow turn,0,0.35,2800")).start();
+                state++;
+                break;
+            case 8:
+                // wait
+                new Thread(new Tasker("pause,500")).start();
+                state++;
+                break;
+            case 10:
+                // turn other way for a second
+                new Thread(new Tasker("slow turn,1,0.35,300")).start();
+                state++;
+                break;
+            case 12:
+                // wait
+                new Thread(new Tasker("pause,501")).start();
+                state++;
+                break;
+            case 14:
+                // bring up base grabby claw
+                //new Thread(new Tasker("BP claw up,750")).start();
+                new Thread(new Tasker("BP claw up,1200")).start();
+                state++;
+                break;
+            case 16:
+                // drive backwards
+                new Thread(new Tasker("drive,0,-1,700")).start();
+                state++;
                 break;
             default:
                 //pause for a bit
-                new Thread(new Tasker("pause,1000"));
+                //new Thread(new Tasker("pause,1000")).start();
                 //state++;
                 break;
         }
-
+        //new Thread(new Tasker("pause,1000")).start();
         telemetry.addData("runtime", runtime.toString() + " seconds");
 
         if (telemetryMsg != null) {
@@ -106,13 +126,49 @@ public class AutoCode extends OpMode {
                     telemetryMsg = "turning with direction " + direction;
                     bot.mech_rotate(direction);
                     pause(Long.parseLong(split[2]));
+                    bot.stopDriving();
                     telemetryMsg = "done rotating";
+                    break;
+                case "slow turn":
+                    // turn slowly than turn
+                    int directionn = Integer.parseInt(split[1]);
+                    telemetryMsg = "turning slowly with direction " + directionn;
+                    bot.mech_rotate(directionn, Double.parseDouble(split[2]));
+                    pause(Long.parseLong(split[3]));
+                    bot.stopDriving();
+                    telemetryMsg = "done turning slowly";
                     break;
                 case "pause":
                     //wait for a bit
                     telemetryMsg = "waiting for a bit";
                     pause(Long.parseLong(split[1]));
                     telemetryMsg = "done waiting";
+                    break;
+                case "open claw":
+                    // Open the claw
+                    telemetryMsg = "Opening the claw";
+                    bot.aux_claw3_direct(1);
+                    pause(Long.parseLong(split[2]));
+                    //bot.aux_claw3(0);
+                    telemetryMsg = "Claw is open";
+                    break;
+                case "close claw":
+                    // Close the claw
+                    telemetryMsg = "Closing the claw";
+                    bot.aux_claw3_direct(0);
+                    pause(Long.parseLong(split[2]));
+                    //bot.aux_claw3(0);
+                    telemetryMsg = "Claw is closed";
+                    break;
+                case "BP claw up":
+                    bot.aux_claw4(1);
+                    pause(Long.parseLong(split[1]));
+                    bot.aux_claw4(0);
+                    break;
+                case "BP claw down":
+                    bot.aux_claw4(-1);
+                    pause(Long.parseLong(split[1]));
+                    bot.aux_claw4(0);
                     break;
                 default:
                     telemetryMsg = split[0] + " is either spelled wrong or hasn't been implemented yet";
