+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TeleOp-Linear-SlowDrive.java
@@ -1,7 +1,7 @@
 /*
 Apex Robotics FTC Team 3916: Main TeleOp for SkyStone season (2019-2020)
 
-Uses a Mechanum-style drivetrain for movement.
+Uses a Mecanum-style drivetrain for movement.
  */
 
 package org.firstinspires.ftc.teamcode;
@@ -9,17 +9,15 @@ package org.firstinspires.ftc.teamcode;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-import org.firstinspires.ftc.robotcore.external.Telemetry;
-
-@TeleOp(name="Linear Slide", group="Apex Robotics 3916")
+@TeleOp(name="Slow Drive", group="Apex Robotics 3916")
 //@Disabled
-public class LinearSlideTeleOp extends OpMode {
+public class SlowDriveTeleOp extends OpMode {
 
     private Robot bot = new Robot();
 
     @Override
     public void init() {
-        bot.init(hardwareMap, Robot.DriveType.Mechanum);
+        bot.init(hardwareMap, Robot.DriveType.Mecanum);
     }
 
     @Override
@@ -38,31 +36,17 @@ public class LinearSlideTeleOp extends OpMode {
             telemetry.addData("Status", "Rotating Clockwise");
             telemetry.update();
         } else {
-            double fx = gamepad1.left_stick_x;
-            double fy = gamepad1.left_stick_y;
-            double sx = gamepad1.right_stick_x;
-            double sy = gamepad1.right_stick_y;
-            String teleStr = "";
+            double sx = gamepad1.left_stick_x;
+            double sy = gamepad1.left_stick_y;
             double slowMult = 0.3;
-            if (Math.abs(sx) > STICK_DEAD_ZONE || Math.abs(sy) > STICK_DEAD_ZONE) {
-                if (Math.abs(sy) > STICK_DEAD_ZONE) {
-                    y = sy * slowMult;
-                }
-                if (Math.abs(sx) > STICK_DEAD_ZONE) {
-                    x = sx * slowMult;
-                }
-                teleStr = "SLOW";
-            } else {
-                if (Math.abs(fy) > STICK_DEAD_ZONE) {
-                    y = fy;
-                }
-                if (Math.abs(fx) > STICK_DEAD_ZONE) {
-                    x = fx;
-                }
-                teleStr = "FAST";
+            if (Math.abs(sy) > STICK_DEAD_ZONE) {
+                y = sy * slowMult;
+            }
+            if (Math.abs(sx) > STICK_DEAD_ZONE) {
+                x = sx * slowMult;
             }
             MechPower pwr = bot.mech_drive(x, y);
-            telemetry.addData("Drive Stat", teleStr + pwr.toString());
+            telemetry.addData("Drive Stat", pwr.toString());
             telemetry.update();
         }
         String message = "Wack nothing's happening...";
@@ -75,10 +59,10 @@ public class LinearSlideTeleOp extends OpMode {
         }
         //stage 2 lift
         if (gamepad2.dpad_up) {
-            bot.aux_claw2(-1); // reversed because servo is backwards see \/
+            bot.aux_claw2(1);
             message = "STAGE 2 LIFT POWER 1";
         } else if (gamepad2.dpad_down) {
-            bot.aux_claw2(1); //reversed because servo is backwards see /\
+            bot.aux_claw2(-1);
             message = "STAGE 2 LIFT POWER -1";
         } else {
             bot.aux_claw2(0);
@@ -93,18 +77,5 @@ public class LinearSlideTeleOp extends OpMode {
             message = "CLAW LIFT POWER 0";
         }
         telemetry.addData("Claw", message);
-
-        // grabber of the build plate
-        if (gamepad2.left_bumper) {
-            bot.aux_claw4(-1);
-            telemetry.addData("Buildplate Grabber", "Power at -1");
-            telemetry.update();
-        } else if (gamepad2.right_bumper) {
-            bot.aux_claw4(1);
-            telemetry.addData("Buildplate Grabber", "Power at 1");
-            telemetry.update();
-        } else {
-            bot.aux_claw4(0);
-        }
     }
 }
