+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TestAuto.java
@@ -1,21 +1,29 @@
+
 package org.firstinspires.ftc.teamcode;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-@Autonomous(name = "Blue Left Autonomous", group = "Apex Robotics 3916")
-public class BlueLeftAuto extends OpMode {
+import org.firstinspires.ftc.robotcore.external.navigation.DistanceUnit;
+
+import java.util.Arrays;
+import java.util.ArrayList;
+import java.util.List;
+
+@Autonomous(name = "Test Autonomous", group = "Apex Robotics 3916")
+public class AutoCode extends OpMode {
     private Robot bot = new Robot();
     private ElapsedTime runtime = new ElapsedTime();
     private String timeOfCompletion;
 
     private int state = 0;
+    public String AutoDone = "no";
     private String telemetryMsg;
 
     @Override
     public void init() {
-        bot.init(hardwareMap, Robot.DriveType.Mechanum);
+        bot.init(hardwareMap, Robot.DriveType.Mecanum);
     }
 
     @Override
@@ -30,35 +38,36 @@ public class BlueLeftAuto extends OpMode {
 
     @Override
     public void loop() {
+        if (AutoDone == "no"){
+            //enter code below, use
+
+            //enter code above
+            AutoDone = "yes";
+
+        }
+        else{
+            //pause for a bit
+            new Thread(new Tasker("pause,1000"));
+        }
         switch (state) {
             // even numbers for performing an action while odd values for state are used for waiting, seen in default case
             case 0:
                 // drive forward
-                new Thread(new Tasker("drive,0,0.5,1500")).start();
+                new Thread(new Tasker("drive,0,1,3000")).start();
                 state++;
 
                 break;
             case 2:
-                // turn left
-                new Thread(new Tasker("turn,0,500")).start();
+                //turn for a bit
+                new Thread(new Tasker("turn,0,1000"));
                 state++;
                 break;
             case 4:
-                // drive forward
-                new Thread(new Tasker("drive,0,0.5,3000")).start();
+                //go forward but sideways
+                new Thread(new Tasker("drive,0,1,3000"));
                 state++;
                 break;
             case 6:
-                // turn right
-                new Thread(new Tasker("turn,1,500")).start();
-                state++;
-                break;
-            case 8:
-                // drive forward
-                new Thread(new Tasker("drive,0,0.5,500")).start();
-                state++;
-                break;
-            case 10:
                 // autonomous is done as robot has parked
                 telemetryMsg = "autonomous completed in " + runtime.toString() + " seconds";
                 break;
@@ -79,6 +88,7 @@ public class BlueLeftAuto extends OpMode {
         telemetry.update();
     }
 
+    //This class is basically a set of functions that you can use in auto
     private class Tasker implements Runnable {
         private String task;
         private String orientation;
